// txc-front-web-toc/src/app/pages/merchants/models/acceptance-loop-details.model.ts
export interface AcceptanceLoopDetails {    
        acceptanceLoopDetailId: number;
        acceptanceLoopId: number;
        merchantId: number;
        shopId: number;
        status: boolean;
        createdBy: string;
        createdOn: string;
        lastModifiedBy: string;
        lastModifiedOn: string;
        action: number;      
}